+++ test/sigil_test.exs
@@ -3,10 +3,10 @@ defmodule MDEx.SigilTest do
   import MDEx.Sigil
 
   test "defaults to HTML" do
-    assert ~MD|# Hello| == "<h1>Hello</h1>\n"
+    assert ~M|# Hello| == "<h1>Hello</h1>\n"
   end
 
   test "AST modifier" do
-    assert ~MD|# Hello|a == [{"document", [], [{"heading", [{"level", 1}, {"setext", false}], ["Hello"]}]}]
+    assert ~M|# Hello|a == [{"document", [], [{"heading", [{"level", 1}, {"setext", false}], ["Hello"]}]}]
   end
 end