+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.2.0-dev
 
 ### Enhancements
-  * Added `~MD` sigil with `a` (AST) modifier (defaults to HTML without the modifier)
+  * Added `~M` sigil with `a` (AST) modifier (defaults to HTML without the modifier)
   * Added `parse_document/1` and `parse_document/2` to parse Markdown to AST
   * Added `traverse_and_update/2` and `attribute/2` to manipulate AST
   * Added `to_html!/1` and `to_html!/2`